- fixed a bug with integer conversion in srUtils.c - changed some lib functions to work on long instead of int to care for 64     bit platforms (just to be on the save side)

+++ srUtils.h
@@ -44,7 +44,7 @@
  *
  * \param iToConv The integer to be converted.
  */
-rsRetVal srUtilItoA(char *pBuf, int iLenBuf, int iToConv);
+rsRetVal srUtilItoA(char *pBuf, int iLenBuf, long iToConv);
 
 /**
  * A method to duplicate a string for which the length is known.
